+++ src/common/interfaces/measure.ts
