--- src/editor/ActionSidebar.tsx
@@ -1,9 +1,0 @@
-import Tabs from "./Tabs";
-
-export default function ActionSidebar() {
-  return (
-    <div className="h-full flex flex-row">
-      <Tabs />
-    </div>
-  );
-}
