--- src/editor/defaultStyles.ts
@@ -1,38 +1,0 @@
-const defaultStyles = {
-  h1: {
-    fontSize: 48,
-    color: "red",
-    marginBottom: 16,
-  },
-  h2: {
-    fontSize: 32,
-    color: "red",
-    marginBottom: 16,
-  },
-  h3: {
-    fontSize: 24,
-    color: "red",
-    marginBottom: 16,
-  },
-  h4: {
-    fontSize: 20,
-    color: "red",
-    marginBottom: 16,
-  },
-  h5: {
-    fontSize: 16,
-    color: "red",
-    marginBottom: 16,
-  },
-  h6: {
-    fontSize: 12,
-    color: "red",
-    marginBottom: 16,
-  },
-  p: {
-    fontSize: 14,
-    marginBottom: 16,
-  },
-};
-
-export default defaultStyles;
