--- src/editor/index.tsx
@@ -1,3 +1,0 @@
-import Editor from "./Editor";
-
-export default Editor;
