+++ src/pdf/pdf.ts
@@ -1,6 +1,6 @@
 import jsPDF from "jspdf";
-import { Box } from "../common/box";
-import { Coord } from "../common/types";
+import { Box } from "../common/classes/box";
+import { Coord } from "../common/interfaces/measure";
 
 export interface LineOptions {
   color?: string;
